Rename BreezNotificationService to BreezNotificationHelper
